+++ packages/jaspr_bootstrap/example/lib/pages/first_page.dart
@@ -1,6 +1,6 @@
 import 'package:jaspr/jaspr.dart';
 import 'package:jaspr/ui.dart';
-import 'package:jaspr_ui/bootstrap.dart';
+import 'package:jaspr_bootstrap/all.dart';
 
 class FirstPage extends StatelessComponent {
   @override
