+++ packages/jaspr_bootstrap/lib/core.dart
@@ -1,0 +1,5 @@
+library core;
+
+export 'src/core/edge_insets.dart';
+export 'src/core/enums.dart';
+export 'src/core/border.dart';
