+++ packages/jaspr_bootstrap/lib/src/core/enums.dart
