+++ packages/jaspr_bootstrap/pubspec.yaml
@@ -1,4 +1,4 @@
-name: jaspr_ui
+name: jaspr_bootstrap
 description: Bootstrap components for jaspr.
 version: 0.1.0
