--- packages/jaspr_ui/README.md
@@ -1,1 +1,0 @@
-# UI components for Jaspr
