--- packages/jaspr_ui/lib/bootstrap.dart
@@ -1,14 +1,0 @@
-library bootstrap;
-
-export 'src/bootstrap/components/carousel.dart';
-export 'src/bootstrap/elements/button.dart';
-export 'src/bootstrap/layout/container.dart';
-export 'src/bootstrap/layout/column.dart';
-export 'src/bootstrap/layout/row.dart';
-export 'src/bootstrap/others/center.dart';
-export 'src/bootstrap/others/size.dart';
-export 'src/bootstrap/edge_insets.dart';
-export 'src/bootstrap/enums.dart';
-export 'src/bootstrap/border.dart';
-
-
